requirements.txt
rdkit-pypi
scikit-learn
pandas
numpy
matplotlib
tqdm
openai
joblib
